Big refactoring.
Split the content into three files: words.v for modeling the problem on words,
walks.v for the interpretation in terms of trajectories on a grid and monad.v
for the technical device used to implement and reason about the bijection.

The invariant for the transformation A -> B is now stated in terms of an
auxiliary code which uses so called 'ghost' variables in order to count
the Dyck words and free occurences processed so far when reading a word.